Fixed a couple import bugs and one apparently unrelated bug.

+++ src/facer/models/daily_average.py
@@ -16,7 +16,7 @@
 
 import numpy as np
 from datetime import timedelta
-from facer.models.full_model import Model
+from .full_model import Model
 
 class DailyAverage(object):
     def __init__(self, phi_d, f_107, day, hemisphere, **kwargs):
@@ -40,7 +40,7 @@ class DailyAverage(object):
         self.ut = {}
         j_totals = []
 
-        for hour in np.arange(24):
+        for hour in np.arange(24, dtype=int):
             self.ut[hour] = Model(phi_d, phi_d, f_107, day + timedelta(hours=hour), hemisphere, **kwargs)
             j_totals.append(self.ut[hour].j_total())
